# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username: string;
  /** @nullable */
  profileImageUrl?: string | null;
  bestScore: number;
  prestigeLevel: number;
  totalRuns: number;
}

export interface LeaderboardResponse {
  entries: LeaderboardEntry[];
}

export interface ScorePayload {
  /** @minimum 0 */
  bestScore: number;
  /** @minimum 0 */
  totalRuns: number;
  /** @minimum 0 */
  cigarettesSmoked: number;
  /** @minimum 0 */
  prestigeLevel: number;
}

export interface ScoreResult {
  bestScore: number;
  updated: boolean;
}

export interface PrestigeResponse {
  prestigeLevel: number;
  multiplier: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};