# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scores:
    post:
      tags: [game]
      operationId: submitScore
      summary: Submit or update a player's score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScorePayload'
      responses:
        '200':
          description: Score recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreResult'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /prestige:
    post:
      tags: [game]
      operationId: performPrestige
      summary: Perform a prestige rebirth
      responses:
        '200':
          description: Prestige performed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrestigeResponse'
        '400':
          description: Prestige requirements not met.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ScorePayload:
      type: object
      required: [bestScore, totalRuns, cigarettesSmoked, prestigeLevel]
      properties:
        bestScore:
          type: integer
          minimum: 0
        totalRuns:
          type: integer
          minimum: 0
        cigarettesSmoked:
          type: integer
          minimum: 0
        prestigeLevel:
          type: integer
          minimum: 0
    ScoreResult:
      type: object
      required: [bestScore, updated]
      properties:
        bestScore:
          type: integer
        updated:
          type: boolean
    PrestigeResponse:
      type: object
      required: [prestigeLevel, multiplier]
      properties:
        prestigeLevel:
          type: integer
        multiplier:
          type: number