# lib/api-zod/src/generated/types/scorePayload.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScorePayload {
  /** @minimum 0 */
  bestScore: number;
  /** @minimum 0 */
  totalRuns: number;
  /** @minimum 0 */
  cigarettesSmoked: number;
  /** @minimum 0 */
  prestigeLevel: number;
}